+++ util/CHANGELOG.md
@@ -5,6 +5,7 @@ Refer to https://keepachangelog.com/en/1.0.0/ for how to maintain changelog<br /
 
 ## [Unreleased]
 
+-   performance: don't pad terminal in production
 -   added `getEnvVar` cross-platform util for getting an environment variable
 
 ## [0.3.1] - 03/05/2023