+++ util/CHANGELOG.md
@@ -5,6 +5,10 @@ Refer to https://keepachangelog.com/en/1.0.0/ for how to maintain changelog<br /
 
 ## [Unreleased]
 
+-
+
+## [0.4.0] - 10/06/2023
+
 -   add process name prefix to logging
 -   performance: don't pad terminal in production
 -   added `getEnvVar` cross-platform util for getting an environment variable